Change type of result to double

+++ src/main/java/Arithmetic.java
@@ -1,7 +1,6 @@
 import Operations.AllOperations;
 import Operations.Operation;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Stack;
 
@@ -9,15 +8,15 @@ import java.util.Stack;
 public class Arithmetic {
     private final List<Operation> operationsList = AllOperations.getAllOperations().getOperationsList();
 
-    private Stack<BigDecimal> stackResult = new Stack<>();
+    private Stack<Double> stackResult = new Stack<>();
 
 
-    public BigDecimal calculation(String expression) {
+    public Double calculation(String expression) {
         Parser parser = new Parser();
         Stack<String> stack = parser.parsing(expression);
         for (String s : stack) {
             try {
-                stackResult.push(new BigDecimal(s));
+                stackResult.push(new Double(s));
             } catch (NumberFormatException e) {
                 for (Operation operation : operationsList) {
                     if (s.equals(operation.getOperator())) {
